bulk update
1. Customerlogin is een class waarbij een gebruiker kan inloggen of een account kan aanmaken
2. email is een class waarin er een email word verzonden naar de gebruiker wanneer inloggegevens zijn vergeten
3. gerechten is een class die eerst als functie in de main stond (opzoeken van gerechten)
4. JsonClass maakt een json file aan van de inloggegevens van een gebruiker
5. JsonClassGerechten hoort een json file te maken van de gerechten (word nog aan gewerkt)
6. JsonClassReservering maakt een json file aan van de reservering van een gebruiker
7. Op de main class staat er nu veel minder code en worden er andere classen aangeroepen
8. menu is een class die eerst als functie in de main stond (opzoeken van het menu van een dag)
9. reservering is een class waarin de gebruiker een reservering kan maken en dit ook kan wijzigen
10. Restaurant class wordt nog aan gewerkt
