+++ Restaurant Application/JsonClassGerechten.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using System.Collections.Generic;
+
+
+public class JsonClassGerechten
+{
+    public int a { get; set; }
+}
