+++ Restaurant Application/Restaurant Application.csproj
@@ -6,4 +6,8 @@
     <RootNamespace>Restaurant_Application</RootNamespace>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+  </ItemGroup>
+
 </Project>
