+++ Restaurant Application/Restaurant.cs
@@ -1,0 +1,14 @@
+﻿
+using System;
+using System.Net;
+using System.Net.Mail;
+
+public class Restaurant
+{
+    public void restaurantFunc()
+    {
+        Console.WriteLine("Restaurant groep 1");
+        Console.WriteLine("het adres is wijnhaven 107");
+        Console.WriteLine("Ons restaurant heeft een capaciteit van 50 mensen");
+    }
+}
